fix(datetime): try catch to if else block
