+++ packages/components/date/index.ts
@@ -22,3 +22,4 @@ export {
 
 export * from './xm-datetime-control';
 export * from './transform-date-string-codec.service';
+export * from './shared/parse-time';
